feat: vote 투표 방식 수정
